match tag and branch in deploy
circleci is soo kind it wrongly sets the environment variable. even though I'm building the tag, it sets the `CIRCLE_BRANCH` env.

So I added the branch to the tag: `$CIRCLE_TAG$CIRCLE_BRANCH`

I could have replaced tag with branch but if they fix in the future, our code will be compatible with that one as well.